requirements.txt
absl-py==1.0.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
cycler==0.11.0
docopt==0.6.2
fonttools==4.28.3
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
gym
idna==3.3
imageio==2.13.1
importlib-metadata==4.8.2
iniconfig==1.1.1
jsonpickle==1.5.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
munch==2.5.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
probscale==0.2.5
protobuf==3.19.1
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacred==0.8.2
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
snakeviz==2.1.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
toml==0.10.2
tomli==1.2.2
torch
torchvision
torchaudio
tornado==6.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
wandb==0.13.8